requirements.txt
streamlit
pandas
lida
Pillow